refactor(capture): CameraViewer section 생성
