+++ src/features/capture/components/CameraViewer.tsx
@@ -1,0 +1,29 @@
+import { CameraView } from "expo-camera";
+import { RefObject } from "react";
+import Animated from "react-native-reanimated";
+import { useZoom } from "../hooks";
+import { GestureDetector } from "react-native-gesture-handler";
+
+type Props = {
+  cameraRef: RefObject<CameraView | null>;
+};
+
+const AnimatedCameraView = Animated.createAnimatedComponent(CameraView);
+
+const CameraViewer = ({ cameraRef }: Props) => {
+  const { pinchGesture, zoomAnimateProps } = useZoom();
+
+  return (
+    <GestureDetector gesture={pinchGesture}>
+      <AnimatedCameraView
+        ref={cameraRef}
+        className="flex-1"
+        facing="back"
+        animateShutter
+        animatedProps={zoomAnimateProps}
+      />
+    </GestureDetector>
+  );
+};
+
+export default CameraViewer;
